requirements.txt
discord.py
asyncpg
python-dotenv
discord-pretty-help
disputils
aiohttp-requests
statcord.py
dblpy
patreon
humanize
jishaku
quart
discord-ext-ipc
Quart-Discord
